ID-8 Split tests into dedicated files
